requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
deprecated==1.2.18
et-xmlfile==2.0.0
flask==3.1.1
flask-bcrypt==1.0.1
flask-limiter==3.12
flask-login==0.6.3
flask-sqlalchemy==3.1.1
flask-wtf==1.2.2
greenlet==3.2.4
gunicorn==23.0.0
itsdangerous==2.2.0
jinja2==3.1.6
limits==5.5.0
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
openpyxl==3.1.5
ordered-set==4.1.0
packaging==25.0
pygments==2.19.2
redis==6.4.0
rich==13.9.4
sqlalchemy==2.0.42
typing-extensions==4.14.1
werkzeug==3.1.3
wrapt==1.17.2
wtforms==3.2.1
python-dotenv==1.2.1
